requirements.txt
reportlab
pillow
pandas
pdf2image
